Tidy mnemonic module's structure
Move cmd_handler outside of mods.rs
Construct exported file's in FileIo::new().
