+++ src/gg20/mnemonic/bip39_bindings.rs
@@ -5,11 +5,10 @@
 //!
 //! Zeroization:
 //!   All functions that accept and/or return structs that implement zeroization:
-//!   [super::Password], [super::Entropy], [bip39::Mnemonic], [bip39::Seed]
+//!   [crate::gg20::Password], [crate::gg20::Entropy], [bip39::Mnemonic], [bip39::Seed]
 
 use super::error::bip39::{Bip39Error::*, Bip39Result};
-use super::Entropy;
-use crate::gg20::types::Password;
+use crate::gg20::types::{Entropy, Password};
 use bip39::{Language, Mnemonic, Seed};
 
 // TODO: we can enrich the API so that users can decide which language they want to use
